Refactored constants to use real integral types to better support custom storage types

+++ tests/suites/linear_imperial.cpp
@@ -7,59 +7,55 @@
 namespace units = ::JadeMatrix::units;
 
 
+// Tests use `long double` values as these are only checking for numerically
+// correct conversions (not type semantics preservation)
+
 TEST_CASE( "feet to inches" )
 {
-    units::feet< double > feet{ 7 };
-    units::inches< double > inches{ feet };
-    REQUIRE( static_cast< double >( inches ) == 84.0 );
+    units::feet< long double > feet{ 7 };
+    units::inches< long double > inches{ feet };
+    REQUIRE( static_cast< long double >( inches ) == 84.0L );
 }
-
 TEST_CASE( "yards to feet" )
 {
-    units::yards< double > yards{ 7 };
-    units::feet< double > feet{ yards };
-    REQUIRE( static_cast< double >( feet ) == 21.0 );
+    units::yards< long double > yards{ 7 };
+    units::feet< long double > feet{ yards };
+    REQUIRE( static_cast< long double >( feet ) == 21.0L );
 }
-
 TEST_CASE( "yards to inches" )
 {
-    units::yards< double > yards{ 7 };
-    units::inches< double > inches{ yards };
-    REQUIRE( static_cast< double >( inches ) == 252.0 );
+    units::yards< long double > yards{ 7 };
+    units::inches< long double > inches{ yards };
+    REQUIRE( static_cast< long double >( inches ) == 252.0L );
 }
-
 TEST_CASE( "miles to feet" )
 {
-    units::miles< double > miles{ 7 };
-    units::feet< double > feet{ miles };
-    REQUIRE( static_cast< double >( feet ) == 36960.0 );
+    units::miles< long double > miles{ 7 };
+    units::feet< long double > feet{ miles };
+    REQUIRE( static_cast< long double >( feet ) == 36960.0L );
 }
-
 
 TEST_CASE( "inches to feet" )
 {
-    units::inches< double > inches{ 7 };
-    units::feet< double > feet{ inches };
-    REQUIRE( static_cast< double >( feet ) == 7.0 / 12.0 );
+    units::inches< long double > inches{ 7 };
+    units::feet< long double > feet{ inches };
+    REQUIRE( static_cast< long double >( feet ) == 7.0L / 12.0L );
 }
-
 TEST_CASE( "feet to yards" )
 {
-    units::feet< double > feet{ 7 };
-    units::yards< double > yards{ feet };
-    REQUIRE( static_cast< double >( yards ) == 7.0 / 3.0 );
+    units::feet< long double > feet{ 7 };
+    units::yards< long double > yards{ feet };
+    REQUIRE( static_cast< long double >( yards ) == 7.0L / 3.0L );
 }
-
 TEST_CASE( "inches to yards" )
 {
-    units::inches< double > inches{ 7 };
-    units::yards< double > yards{ inches };
-    REQUIRE( static_cast< double >( yards ) == 7.0 / 36.0 );
+    units::inches< long double > inches{ 7 };
+    units::yards< long double > yards{ inches };
+    REQUIRE( static_cast< long double >( yards ) == 7.0L / 36.0L );
 }
-
 TEST_CASE( "feet to miles" )
 {
-    units::feet< double > feet{ 7 };
-    units::miles< double > miles{ feet };
-    REQUIRE( static_cast< double >( miles ) == 7.0 / 5280.0 );
+    units::feet< long double > feet{ 7 };
+    units::miles< long double > miles{ feet };
+    REQUIRE( static_cast< long double >( miles ) == 7.0L / 5280.0L );
 }
